chore: rename folder to match key
